Redux implementation of auth
