--- new_client/src/components/Error.css
@@ -1,4 +1,0 @@
-.errorMsg {
-    color: #cc0000;
-    margin-bottom: 10px;
-}
