+++ new_client/src/components/Header.js
@@ -13,7 +13,7 @@ export default function Header() {
         <Link to="/"><i className="fa fa-fw fa-info"></i>EVENTS</Link>
         <Link to="/glimpse"><i className="fa fa-fw fa-info"></i>GLIMPSES</Link>
         <Link to="/blogs"><i className="fa fa-fw fa-info"></i>BLOGS</Link>
-        <Link to="/auth"><i className="fa fa-fw fa-info"></i> SIGN UP</Link>
+        <Link to="/signup"><i className="fa fa-fw fa-info"></i> SIGN UP</Link>
 
       </div>
     </div>
