+++ new_client/src/components/auth/Signin.jsx
@@ -1,12 +1,20 @@
-import React from 'react';
-import { Container, Paper, Grid, TextField, Button } from '@material-ui/core';
+import React, {useState, useEffect} from 'react';
+import { Container, Paper, Grid, TextField, Button, Typography, Link } from '@material-ui/core';
 import "./Error.css";
+import { connect } from "react-redux";
+import { login } from "../../actions/authActions";
+import { clearErrors } from "../../actions/errorActions";
 
-export default function Signin() {
-	const [form, updateForm] = React.useState({
-		username: "",
-		password: "",
-	});
+function Signin(props) {
+
+	const { isAuthenticated, error, login, clearErrors } = props; 
+
+	const [username, setUsername] = useState('');
+	const [password, setPassword] = useState('');
+	const [msg, setMsg] = useState(null);
+
+	const handleUsername= (e) => setUsername(e.target.value);
+	const handlePassword= (e) => setPassword(e.target.value);
 
 	const [errors, updateErrors] = React.useState({
 		username: "",
@@ -15,7 +23,7 @@ export default function Signin() {
 
 	function isFormValid() {
 		let formIsValid = true;
-		if (!form.username) {
+		if (!username) {
 			formIsValid = false;
 			updateErrors(prevErrors => ({
 				...prevErrors,
@@ -23,7 +31,7 @@ export default function Signin() {
 			}));
 		}
 
-		if (!form.password) {
+		if (!password) {
 			formIsValid = false;
 			updateErrors(prevErrors => ({
 				...prevErrors,
@@ -35,39 +43,25 @@ export default function Signin() {
 
 	}
 
-
-	function handleChange(event) {
-		const { name, value } = event.target;
-		updateForm(prevDetails => {
-			return (
-				{
-					...prevDetails,
-					[name]: value
-				}
-			);
-		});
-	}
-
 	function handleClick(event) {
+		event.preventDefault();
 		if (isFormValid()) {
-			event.preventDefault();
-			fetch(process.env.REACT_APP_API_LOGIN, {
-				method: "POST",
-				mode: "cors",
-				headers: {
-					"Content-Type": "application/json",
-				},
-				body: JSON.stringify(form),
-			})
-				.then((response) => response.json())
-				.catch((err) => {
-					console.log(err);
-				});
+			const user = { username, password };
+			login(user);
 		}
 		else {
 			alert("There are errors in your form !");
 		}
 	}
+
+	useEffect(() => {
+		if (error.id === 'LOGIN_FAIL') {
+			setMsg(error.msg.msg);
+		}
+		else {
+			setMsg(null);
+		}
+	})
 
 	return (
 		<Container maxWidth="sm">
@@ -81,8 +75,7 @@ export default function Signin() {
 							name="username"
 							type="text"
 							placeholder="Username"
-							onChange={handleChange}
-							value={form.username}
+							onChange={handleUsername}
 						/>
 						<div className="errorMsg">{errors.username}</div>
 					</Grid>
@@ -94,8 +87,7 @@ export default function Signin() {
 							name="password"
 							type="password"
 							placeholder="Password"
-							onChange={handleChange}
-							value={form.password}
+							onChange={handlePassword}
 						/>
 						<div className="errorMsg">{errors.password}</div>
 					</Grid>
@@ -109,9 +101,22 @@ export default function Signin() {
 							Submit
                 </Button>
 					</Grid>
+
+					<Grid item xs={12}>
+						<Typography>
+							New User ? <Link to="/signup">Sign Up</Link> instead.
+						</Typography>
+					</Grid>
 				</Grid>
 			</Paper>
 		</Container>
 
 	)
 }
+
+const mapStateToProps = (state) => ({
+	isAuthenticated : state.auth.isAuthenticated,
+	error : state.error
+});
+
+export default connect(mapStateToProps, { login, clearErrors })(Signin);
