+++ new_client/src/reducers/authReducer.js
@@ -29,6 +29,7 @@ export default function (state = initialState, action) {
       };
     case LOGIN_SUCCESS:
     case REGISTER_SUCCESS:
+      console.log(action.payload)
       localStorage.setItem("token", action.payload.token);
       return {
         ...state,
