+++ new_client/src/store/configureStore.js
@@ -1,11 +1,16 @@
-import { createStore } from "redux";
+import { createStore, compose, applyMiddleware } from "redux";
 import rootReducer from "../reducers";
+import thunk from "redux-thunk";
 
 export default function configureStore(initialState) {
+
+  const middleware = [thunk];
+  const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
+
   const store = createStore(
     rootReducer,
     initialState,
-    window.devToolsExtension ? window.devToolsExtension() : undefined
+    composeEnhancers(applyMiddleware(...middleware))
   );
 
   if (module.hot) {
